backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
numpy==1.26.4
pillow==10.3.0
pycparser==2.21
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
